+++ quiche/src/recovery/gcongestion/bbr2/network_model.rs
@@ -32,9 +32,9 @@ use std::ops::Add;
 use std::time::Duration;
 use std::time::Instant;
 
-use crate::recovery::gcongestion::bandwidth::Bandwidth;
 use crate::recovery::gcongestion::bbr::BandwidthSampler;
 use crate::recovery::gcongestion::bbr2::Params;
+use crate::recovery::gcongestion::Bandwidth;
 use crate::recovery::gcongestion::Lost;
 use crate::recovery::rtt::RttStats;
 use crate::recovery::rtt::INITIAL_RTT;
