+++ quiche/src/recovery/mod.rs
@@ -1694,6 +1694,7 @@ mod tests {
     }
 }
 
+mod bandwidth;
 mod congestion;
 mod gcongestion;
 mod rtt;
